refactor: extract helpers.rs from app.rs
Extract 6 free functions from app.rs (2,449 -> 2,296 lines):
- helpers.rs (163 lines): fire_webhook, fire_notification,
  kill_process, dirs_home, chrono_now_iso, create_aggregate_session
